pyproject.toml
[project]
name = "qualificadordeleadia"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.15",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anyio>=4.10.0",
    "appdirs>=1.4.4",
    "asttokens>=3.0.0",
    "asyncpg>=0.30.0",
    "attrs>=25.3.0",
    "backoff>=2.2.1",
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.13.4",
    "blinker>=1.9.0",
    "bs4>=0.0.2",
    "build>=1.3.0",
    "cachetools>=5.5.2",
    "certifi>=2025.8.3",
    "cffi>=1.17.1",
    "charset-normalizer>=3.4.2",
    "chromadb>=1.0.15",
    "click>=8.2.1",
    "colorama>=0.4.6",
    "coloredlogs>=15.0.1",
    "crewai>=0.152.0",
    "cryptography>=45.0.5",
    "dataclasses-json>=0.6.7",
    "decorator>=5.2.1",
    "diskcache>=5.6.3",
    "distro>=1.9.0",
    "dnspython>=2.7.0",
    "docstring_parser>=0.17.0",
    "durationpy>=0.10",
    "email_validator>=2.2.0",
    "et_xmlfile>=2.0.0",
    "executing>=2.2.0",
    "fastapi>=0.116.1",
    "filelock>=3.18.0",
    "Flask>=3.1.1",
    "flatbuffers>=25.2.10",
    "frozenlist>=1.7.0",
    "fsspec>=2025.7.0",
    "google-auth>=2.40.3",
    "googleapis-common-protos>=1.70.0",
    "greenlet>=3.2.3",
    "grpcio>=1.74.0",
    "gunicorn>=23.0.0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httptools>=0.6.4",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.1",
    "httpx-ws>=0.7.2",
    "huggingface-hub>=0.34.3",
    "humanfriendly>=10.0",
    "idna>=3.10",
    "importlib_metadata>=8.7.0",
    "importlib_resources>=6.5.2",
    "instructor>=1.10.0",
    "ipython>=9.4.0",
    "ipython_pygments_lexers>=1.1.1",
    "itsdangerous>=2.2.0",
    "jedi>=0.19.2",
    "Jinja2>=3.1.6",
    "jiter>=0.10.0",
    "joblib>=1.5.1",
    "json5>=0.12.0",
    "json_repair>=0.25.2",
    "jsonpatch>=1.33",
    "jsonpickle>=4.1.1",
    "jsonpointer>=3.0.0",
    "jsonref>=1.1.0",
    "jsonschema>=4.25.0",
    "jsonschema-specifications>=2025.4.1",
    "kubernetes>=33.1.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "langchain-text-splitters>=0.3.9",
    "langgraph>=0.6.3",
    "langgraph-checkpoint>=2.1.1",
    "langgraph-prebuilt>=0.6.3",
    "langgraph-sdk>=0.2.0",
    "langsmith>=0.4.11",
    "litellm>=1.74.3",
    "lmstudio>=1.5.0",
    "markdown-it-py>=3.0.0",
    "MarkupSafe>=3.0.2",
    "marshmallow>=3.26.1",
    "matplotlib-inline>=0.1.7",
    "mdurl>=0.1.2",
    "mmh3>=5.2.0",
    "mpmath>=1.3.0",
    "msgspec>=0.19.0",
    "multidict>=6.6.3",
    "mypy_extensions>=1.1.0",
    "networkx>=3.5",
    "nltk>=3.9.1",
    "numpy>=2.3.2",
    "oauthlib>=3.3.1",
    "onnxruntime>=1.22.0",
    "openai>=1.99.0",
    "openpyxl>=3.1.5",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-otlp-proto-common>=1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.36.0",
    "opentelemetry-exporter-otlp-proto-http>=1.36.0",
    "opentelemetry-proto>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-semantic-conventions>=0.57b0",
    "orjson>=3.11.1",
    "ormsgpack>=1.10.0",
    "overrides>=7.7.0",
    "packaging>=25.0",
    "parso>=0.8.4",
    "pdfminer.six>=20250506",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "portalocker>=2.7.0",
    "posthog>=5.4.0",
    "prompt_toolkit>=3.0.51",
    "propcache>=0.3.2",
    "protobuf>=6.31.1",
    "pure_eval>=0.2.3",
    "pyasn1>=0.6.1",
    "pyasn1_modules>=0.4.2",
    "pybase64>=1.4.2",
    "pycparser>=2.22",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pydantic_core>=2.33.2",
    "Pygments>=2.19.2",
    "PyJWT>=2.10.1",
    "pypdfium2>=4.30.0",
    "PyPika>=0.48.9",
    "pyproject_hooks>=1.2.0",
    "pyreadline3>=3.5.4",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pyvis>=0.3.2",
    "PyYAML>=6.0.2",
    "referencing>=0.36.2",
    "regex>=2025.7.34",
    "reportlab>=4.4.3",
    "requests>=2.32.4",
    "requests-oauthlib>=2.0.0",
    "requests-toolbelt>=1.0.0",
    "rich>=14.1.0",
    "rpds-py>=0.26.0",
    "rsa>=4.9.1",
    "safetensors>=0.6.2",
    "sentencepiece>=0.2.1",
    "shellingham>=1.5.4",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "soupsieve>=2.7",
    "SQLAlchemy>=2.0.42",
    "stack-data>=0.6.3",
    "starlette>=0.47.2",
    "sympy>=1.14.0",
    "tenacity>=9.1.2",
    "textblob>=0.19.0",
    "tiktoken>=0.9.0",
    "tokenizers>=0.21.4",
    "tomli>=2.2.1",
    "tomli_w>=1.2.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "traitlets>=5.14.3",
    "transformers>=4.55.2",
    "typer>=0.16.0",
    "typing-inspect>=0.9.0",
    "typing-inspection>=0.4.1",
    "typing_extensions>=4.14.1",
    "urllib3>=2.5.0",
    "uv>=0.8.4",
    "uvicorn>=0.35.0",
    "watchfiles>=1.1.0",
    "wcwidth>=0.2.13",
    "websocket-client>=1.8.0",
    "websockets>=15.0.1",
    "Werkzeug>=3.1.3",
    "wsproto>=1.2.0",
    "xxhash>=3.5.0",
    "yarl>=1.20.1",
    "zipp>=3.23.0",
    "zstandard>=0.23.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

